2022-5-10   14:49 bug fixed
pxb

+++ service_xiao/src/main/java/ruangong/root/service_xiao/impl/TemplateServiceImpl.java
@@ -88,7 +88,7 @@ public class TemplateServiceImpl extends ServiceImpl<TemplateMapper, Template> i
 
         QueryWrapper<Template> queryWrapper = Wrappers.query();
         queryWrapper.eq("id", id);
-        Integer count = templateMapper.selectCount(queryWrapper);
+        Long count =  (long) templateMapper.selectCount(queryWrapper);
         if (count != 1) {
             ResultUtil.quickSet(
                     result,
